Feature - Downloading File (invoices, images, etc.)

+++ controllers/admin.js
@@ -1,5 +1,6 @@
-const Product = require("../models/product");
 const { validationResult } = require("express-validator");
+const Product = require("../models/product");
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -131,6 +132,7 @@ exports.postEditProduct = (req, res, next) => {
       product.description = updatedDesc;
       /* update image only there is new image is uploaded */
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -159,10 +161,21 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found!"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
     })
-    .catch((err) => console.log(err));
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
+      return next(error);
+    });
 };
